% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the DegreeDayCalc Shiny application (internal helper)}
\usage{
run_app(...)
}
\arguments{
\item{...}{Passed to shiny::runApp().}
}
\description{
Internal helper that launches the bundled Shiny app.
}
\keyword{internal}
